Purges admin post form that allowed staff to post with any username, automagically creating it if it did not exist.
Don't agree with the need for admins to have the ability to post as any username or create users on the fly.  It is confusing for end users that do not request it, and looks very much like a vulnerability.  Those that do need the functionality can use a 3rd party package such as django-hijack or subclass the view and write their own `get_form_class` implementation.